+++ src/networks/types/AVAX.js
